requirements.txt

# This file contains all of our immediate dependencies. Each line also contains the license of the respective dependency.

torch==1.8.1            #BSD 3-Clause
numpy==1.20.2           #BSD 3-Clause
Pillow==8.2.0           #HPND license
torchvision==0.9.1      #BSD 3-Clause
inputs==0.5             #BSD 3-Clause
requests==2.25.1        #Apache 2.0
PySide6==6.0.3          #GNU LGPLv3